raycasting functions are 25 lines or less
