+++ Makefile
@@ -6,7 +6,7 @@
 #    By: sniemela <sniemela@student.hive.fi>        +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2025/02/19 13:52:26 by hpirkola          #+#    #+#              #
-#    Updated: 2025/04/07 12:45:47 by hpirkola         ###   ########.fr        #
+#    Updated: 2025/04/07 17:35:22 by sniemela         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -38,6 +38,7 @@ SOURCES = main.c \
 	init_utils_bonus.c \
 	moving_bonus.c \
 	parsing.c \
+	raycasting.c \
 	terminate_n_free_bonus.c \
 	utils.c \
 	validate.c
